Parsing Errors added

+++ HandleRequests/StringManip.hpp
@@ -13,7 +13,7 @@ public:
 };
 
 
-int hex_to_num(std::string hex);
+long long hex_to_num(std::string hex);
 long long stringToNumber(std::string s);
 void exit_error(const char * s);
 void write_str(const char * des , const char *s);
